Iniital commit
The idea here is (well the idea is just to build a library for fun in
this case, let's be honest) to create a dependency-free DNS resolver in
PHP that's easy to test.

There are a few other libraries out there (spatie/dns for example) which
is just fine, but it has at least one symfony requirement, and for
something as simple as DNS resolution I think it should be
dependency-free.

Right now, because a DNS record generally has 4 common properties (host,
TTL, type, class), but specific record types have unique properties, we
have extended interfaces.  Something doesn't feel right about that, but
it will work for now. As more types are added, the design can be
revisited.